Remove nearly all references to orthonormal_basis.h
