+++ bindings/pydrake/math_py.cc
@@ -21,7 +21,6 @@
 #include "drake/math/discrete_algebraic_riccati_equation.h"
 #include "drake/math/discrete_lyapunov_equation.h"
 #include "drake/math/matrix_util.h"
-#include "drake/math/orthonormal_basis.h"
 #include "drake/math/quadratic_form.h"
 #include "drake/math/random_rotation.h"
 #include "drake/math/rigid_transform.h"
