+++ geometry/proximity/BUILD.bazel
@@ -629,7 +629,6 @@ drake_cc_library(
         "//geometry:proximity_engine",
         "//geometry:utilities",
         "//math:geometric_transform",
-        "//math:orthonormal_basis",
     ],
 )
 
@@ -852,8 +851,8 @@ drake_cc_googletest(
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:expect_throws_message",
         "//math:autodiff",
+        "//math:geometric_transform",
         "//math:gradient",
-        "//math:orthonormal_basis",
     ],
 )
 
@@ -865,8 +864,8 @@ drake_cc_googletest(
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:expect_throws_message",
         "//math:autodiff",
+        "//math:geometric_transform",
         "//math:gradient",
-        "//math:orthonormal_basis",
     ],
 )
