+++ math/BUILD.bazel
@@ -152,7 +152,7 @@ drake_cc_library(
 
 drake_cc_library(
     name = "orthonormal_basis",
-    srcs = ["orthonormal_basis.cc"],
+    srcs = [],
     hdrs = ["orthonormal_basis.h"],
     deps = [
         ":geometric_transform",
