--- math/orthonormal_basis.cc
@@ -1,4 +1,0 @@
-// For now, this is an empty .cc file that only serves to confirm
-// orthonormal_basis.h is a stand-alone header.
-
-#include "drake/math/orthonormal_basis.h"
